+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/DamagesService.kt
@@ -35,6 +35,6 @@ class DamagesService(
       }
     )
 
-    return save(reportedAdjudication)
+    return saveToDto(reportedAdjudication)
   }
 }
