+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/ReportedAdjudicationService.kt
@@ -4,42 +4,18 @@ import com.microsoft.applicationinsights.TelemetryClient
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.DraftAdjudicationDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.HearingDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.HearingSummaryDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentDetailsDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentRoleDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentStatementDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceRuleDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.ReportedAdjudicationDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.ReportedDamageDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.ReportedEvidenceDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.ReportedWitnessDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Damage
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.DraftAdjudication
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Evidence
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Hearing
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentDetails
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentRole
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentStatement
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Offence
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedAdjudication
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedAdjudicationStatus
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedDamage
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedEvidence
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedOffence
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedWitness
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Witness
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.gateways.AdjudicationDetailsToPublish
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.gateways.PrisonApiGateway
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.DraftAdjudicationRepository
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.HearingRepository
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.ReportedAdjudicationRepository
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.security.AuthenticationFacade
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.IncidentRoleRuleLookup
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.OffenceCodeLookupService
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.toDto
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
@@ -50,28 +26,26 @@ import javax.transaction.Transactional
 @Transactional
 @Service
 class ReportedAdjudicationService(
-  val draftAdjudicationRepository: DraftAdjudicationRepository,
-  val reportedAdjudicationRepository: ReportedAdjudicationRepository,
+  reportedAdjudicationRepository: ReportedAdjudicationRepository,
   val prisonApiGateway: PrisonApiGateway,
-  val offenceCodeLookupService: OffenceCodeLookupService,
-  val authenticationFacade: AuthenticationFacade,
+  offenceCodeLookupService: OffenceCodeLookupService,
+  authenticationFacade: AuthenticationFacade,
   val telemetryClient: TelemetryClient,
   val hearingRepository: HearingRepository
+) : ReportedAdjudicationBaseService(
+  reportedAdjudicationRepository, offenceCodeLookupService, authenticationFacade
 ) {
   companion object {
     const val TELEMETRY_EVENT = "ReportedAdjudicationStatusEvent"
-    fun throwEntityNotFoundException(id: Long): Nothing =
-      throw EntityNotFoundException("ReportedAdjudication not found for $id")
     fun reportsFrom(startDate: LocalDate): LocalDateTime = startDate.atStartOfDay()
     fun reportsTo(endDate: LocalDate): LocalDateTime = endDate.atTime(LocalTime.MAX)
     fun statuses(status: Optional<ReportedAdjudicationStatus>): List<ReportedAdjudicationStatus> = status.map { listOf(it) }.orElse(ReportedAdjudicationStatus.values().toList())
   }
 
   fun getReportedAdjudicationDetails(adjudicationNumber: Long): ReportedAdjudicationDto {
-    val reportedAdjudication =
-      reportedAdjudicationRepository.findByReportNumber(adjudicationNumber)
+    val reportedAdjudication = findByAdjudicationNumber(adjudicationNumber)
 
-    return reportedAdjudication?.toDto(offenceCodeLookupService) ?: throwEntityNotFoundException(adjudicationNumber)
+    return reportedAdjudication.toDto()
   }
 
   fun getAllReportedAdjudications(agencyId: String, startDate: LocalDate, endDate: LocalDate, status: Optional<ReportedAdjudicationStatus>, pageable: Pageable): Page<ReportedAdjudicationDto> {
@@ -79,7 +53,7 @@ class ReportedAdjudicationService(
       reportedAdjudicationRepository.findByAgencyIdAndDateTimeOfDiscoveryBetweenAndStatusIn(
         agencyId, reportsFrom(startDate), reportsTo(endDate), statuses(status), pageable
       )
-    return reportedAdjudicationsPage.map { it.toDto(offenceCodeLookupService) }
+    return reportedAdjudicationsPage.map { it.toDto() }
   }
 
   fun getMyReportedAdjudications(agencyId: String, startDate: LocalDate, endDate: LocalDate, status: Optional<ReportedAdjudicationStatus>, pageable: Pageable): Page<ReportedAdjudicationDto> {
@@ -89,48 +63,7 @@ class ReportedAdjudicationService(
       reportedAdjudicationRepository.findByCreatedByUserIdAndAgencyIdAndDateTimeOfDiscoveryBetweenAndStatusIn(
         username!!, agencyId, reportsFrom(startDate), reportsTo(endDate), statuses(status), pageable
       )
-    return reportedAdjudicationsPage.map { it.toDto(offenceCodeLookupService) }
-  }
-
-  fun createDraftFromReportedAdjudication(adjudicationNumber: Long): DraftAdjudicationDto {
-    val foundReportedAdjudication =
-      reportedAdjudicationRepository.findByReportNumber(adjudicationNumber)
-
-    val reportedAdjudication = foundReportedAdjudication ?: throwEntityNotFoundException(adjudicationNumber)
-
-    val draftAdjudication = DraftAdjudication(
-      reportNumber = reportedAdjudication.reportNumber,
-      reportByUserId = reportedAdjudication.createdByUserId,
-      prisonerNumber = reportedAdjudication.prisonerNumber,
-      agencyId = reportedAdjudication.agencyId,
-      incidentDetails = IncidentDetails(
-        locationId = reportedAdjudication.locationId,
-        dateTimeOfIncident = reportedAdjudication.dateTimeOfIncident,
-        dateTimeOfDiscovery = reportedAdjudication.dateTimeOfDiscovery,
-        handoverDeadline = reportedAdjudication.handoverDeadline
-      ),
-      incidentRole = IncidentRole(
-        roleCode = reportedAdjudication.incidentRoleCode,
-        associatedPrisonersNumber = reportedAdjudication.incidentRoleAssociatedPrisonersNumber,
-        associatedPrisonersName = reportedAdjudication.incidentRoleAssociatedPrisonersName,
-      ),
-      offenceDetails = toDraftOffence(reportedAdjudication.offenceDetails),
-      incidentStatement = IncidentStatement(
-        statement = reportedAdjudication.statement,
-        completed = true
-      ),
-      isYouthOffender = reportedAdjudication.isYouthOffender,
-      damages = toDraftDamages(reportedAdjudication.damages),
-      evidence = toDraftEvidence(reportedAdjudication.evidence),
-      witnesses = toDraftWitnesses(reportedAdjudication.witnesses),
-      damagesSaved = true,
-      evidenceSaved = true,
-      witnessesSaved = true,
-    )
-
-    return draftAdjudicationRepository
-      .save(draftAdjudication)
-      .toDto(offenceCodeLookupService)
+    return reportedAdjudicationsPage.map { it.toDto() }
   }
 
   fun getAllHearingsByAgencyIdAndDate(agencyId: String, dateOfHearing: LocalDate): List<HearingSummaryDto> {
@@ -145,52 +78,13 @@ class ReportedAdjudicationService(
     return toHearingSummaries(hearings, adjudicationsMap)
   }
 
-  private fun toDraftOffence(offences: MutableList<ReportedOffence>): MutableList<Offence> =
-    offences.map { offence ->
-      Offence(
-        offenceCode = offence.offenceCode,
-        victimPrisonersNumber = offence.victimPrisonersNumber,
-        victimStaffUsername = offence.victimStaffUsername,
-        victimOtherPersonsName = offence.victimOtherPersonsName,
-      )
-    }.toMutableList()
-
-  private fun toDraftDamages(damages: MutableList<ReportedDamage>): MutableList<Damage> =
-    damages.map {
-      Damage(
-        code = it.code,
-        details = it.details,
-        reporter = it.reporter
-      )
-    }.toMutableList()
-
-  private fun toDraftEvidence(evidence: MutableList<ReportedEvidence>): MutableList<Evidence> =
-    evidence.map {
-      Evidence(
-        code = it.code,
-        details = it.details,
-        reporter = it.reporter,
-        identifier = it.identifier
-      )
-    }.toMutableList()
-
-  private fun toDraftWitnesses(witnesses: MutableList<ReportedWitness>): MutableList<Witness> =
-    witnesses.map {
-      Witness(
-        code = it.code,
-        firstName = it.firstName,
-        lastName = it.lastName,
-        reporter = it.reporter
-      )
-    }.toMutableList()
-
   fun setStatus(adjudicationNumber: Long, status: ReportedAdjudicationStatus, statusReason: String? = null, statusDetails: String? = null): ReportedAdjudicationDto {
     val username = if (status == ReportedAdjudicationStatus.AWAITING_REVIEW) null else authenticationFacade.currentUsername
     val reportedAdjudication = reportedAdjudicationRepository.findByReportNumber(adjudicationNumber)
       ?: throw EntityNotFoundException("ReportedAdjudication not found for reported adjudication number $adjudicationNumber")
     val reportedAdjudicationToReturn = reportedAdjudication.let {
       it.transition(to = status, reason = statusReason, details = statusDetails, reviewUserId = username)
-      reportedAdjudicationRepository.save(it).toDto(this.offenceCodeLookupService)
+      saveToDto(it)
     }
     if (status.isAccepted()) {
       saveToPrisonApi(reportedAdjudication)
@@ -260,92 +154,6 @@ class ReportedAdjudicationService(
     return offenceDetails?.mapNotNull { it.victimStaffUsername } ?: emptyList()
   }
 }
-
-fun ReportedAdjudication.toDto(offenceCodeLookupService: OffenceCodeLookupService): ReportedAdjudicationDto = ReportedAdjudicationDto(
-  adjudicationNumber = reportNumber,
-  prisonerNumber = prisonerNumber,
-  bookingId = bookingId,
-  incidentDetails = IncidentDetailsDto(
-    locationId = locationId,
-    dateTimeOfIncident = dateTimeOfIncident,
-    dateTimeOfDiscovery = dateTimeOfDiscovery,
-    handoverDeadline = handoverDeadline
-  ),
-  isYouthOffender = isYouthOffender,
-  incidentRole = IncidentRoleDto(
-    roleCode = incidentRoleCode,
-    offenceRule = IncidentRoleRuleLookup.getOffenceRuleDetails(incidentRoleCode, isYouthOffender),
-    associatedPrisonersNumber = incidentRoleAssociatedPrisonersNumber,
-    associatedPrisonersName = incidentRoleAssociatedPrisonersName,
-  ),
-  offenceDetails = toReportedOffence(offenceDetails, isYouthOffender, offenceCodeLookupService),
-  incidentStatement = IncidentStatementDto(
-    statement = statement,
-    completed = true,
-  ),
-  createdByUserId = createdByUserId!!,
-  createdDateTime = createDateTime!!,
-  reviewedByUserId = reviewUserId,
-  damages = toReportedDamages(damages),
-  evidence = toReportedEvidence(evidence),
-  witnesses = toReportedWitnesses(witnesses),
-  status = status,
-  statusReason = statusReason,
-  statusDetails = statusDetails,
-  hearings = toHearings(hearings),
-)
-
-private fun toReportedOffence(offences: MutableList<ReportedOffence>, isYouthOffender: Boolean, offenceCodeLookupService: OffenceCodeLookupService): List<OffenceDto> =
-  offences.map { offence ->
-    OffenceDto(
-      offenceCode = offence.offenceCode,
-      offenceRule = OffenceRuleDto(
-        paragraphNumber = offenceCodeLookupService.getParagraphNumber(offence.offenceCode, isYouthOffender),
-        paragraphDescription = offenceCodeLookupService.getParagraphDescription(offence.offenceCode, isYouthOffender),
-      ),
-      victimPrisonersNumber = offence.victimPrisonersNumber,
-      victimStaffUsername = offence.victimStaffUsername,
-      victimOtherPersonsName = offence.victimOtherPersonsName,
-    )
-  }.toList()
-
-private fun toReportedDamages(damages: MutableList<ReportedDamage>): List<ReportedDamageDto> =
-  damages.map {
-    ReportedDamageDto(
-      code = it.code,
-      details = it.details,
-      reporter = it.reporter
-    )
-  }.toList()
-
-private fun toReportedEvidence(evidence: MutableList<ReportedEvidence>): List<ReportedEvidenceDto> =
-  evidence.map {
-    ReportedEvidenceDto(
-      code = it.code,
-      identifier = it.identifier,
-      details = it.details,
-      reporter = it.reporter
-    )
-  }.toList()
-
-private fun toReportedWitnesses(witnesses: MutableList<ReportedWitness>): List<ReportedWitnessDto> =
-  witnesses.map {
-    ReportedWitnessDto(
-      code = it.code,
-      firstName = it.firstName,
-      lastName = it.lastName,
-      reporter = it.reporter
-    )
-  }.toList()
-
-private fun toHearings(hearings: MutableList<Hearing>): List<HearingDto> =
-  hearings.map {
-    HearingDto(
-      id = it.id,
-      locationId = it.locationId,
-      dateTimeOfHearing = it.dateTimeOfHearing
-    )
-  }.toList()
 
 private fun toHearingSummaries(hearings: List<Hearing>, adjudications: Map<Long, ReportedAdjudication>): List<HearingSummaryDto> =
   hearings.map {
