+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/WitnessesService.kt
@@ -37,6 +37,6 @@ class WitnessesService(
       }
     )
 
-    return save(reportedAdjudication)
+    return saveToDto(reportedAdjudication)
   }
 }
