+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/EvidenceServiceTest.kt
@@ -69,7 +69,7 @@ class EvidenceServiceTest : ReportedAdjudicationTestBase() {
   }
 
   @Test
-  fun `throws an entity not found if the reported adjudication for the supplied id does not exists`() {
+  override fun `throws an entity not found if the reported adjudication for the supplied id does not exists`() {
     whenever(reportedAdjudicationRepository.findByReportNumber(any())).thenReturn(null)
 
     Assertions.assertThatThrownBy {
